Force tcp connections to close across C/R
Dumping and restoring TCP connections requires that we run as root
inside the container (or have the relevant caps) to allow CRIU to use
iptables or nftables to lock/unlock the network. To avoid that we can
ask CRIU to close any established connections; on restore the program
will see that the connection was closed and handle it like any other
disconnection. In this particular case it tries to reconnect to the
remote hosts and everything works as expected.

Semeru doesn't currently expose a Java property to allow users to
control this, so we'll use CRIU's default conf file instead.