Fix race cond between automation and REST services
The driver script looks for "started in.*Listening on" in stdout to
determine whether or not the REST services have started. In the
Semeru restore script we dump all output that was captured until
the checkpoint, which works if the service was able to reach
listening before being checkpointed, but fails on slow machines
if it didn't. Fix this by continuously dumping the output so that
if the service starts listening after being restored the driver
script will still see the expected output.